Move to use tina instead of .tina
