+++ pages/index.js
@@ -1,7 +1,7 @@
 import { TinaMarkdown } from "tinacms/dist/rich-text";
 import { Layout } from "../components/Layout";
 import { useTina } from "tinacms/dist/react";
-import { client } from "../.tina/__generated__/client";
+import { client } from "../tina/__generated__/client";
 
 export default function Home(props) {
   // data passes though in production mode and data is updated to the sidebar data in edit-mode
