+++ tina/config.js
@@ -1,9 +1,15 @@
 import { defineConfig } from "tinacms";
 
-export const config = defineConfig({
+const branch =
+  process.env.NEXT_PUBLIC_TINA_BRANCH ||
+  process.env.NEXT_PUBLIC_VERCEL_GIT_COMMIT_REF ||
+  process.env.HEAD ||
+  'main';
+
+export default defineConfig({
   clientId: process.env.NEXT_PUBLIC_TINA_CLIENT_ID,
   token: process.env.TINA_TOKEN,
-  branch: "main",
+  branch: branch,
   // Relative to the _root_ of your repo
   localContentPath: "../../demo-content-repo",
   build: {
@@ -63,5 +69,3 @@ export const config = defineConfig({
     ],
   },
 });
-
-export default config;
